pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-speech"
version = "0.5.0"
description = "OpenAI-compatible speech server — any STT/TTS provider, one container"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{ name = "Jeremy Windsor" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "faster-whisper>=1.0.0",
    "huggingface-hub>=0.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "onnxruntime>=1.16.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "wyoming>=1.5.0",
]

[project.optional-dependencies]
moonshine = ["useful-moonshine-onnx>=20251121"]
vosk = ["vosk>=0.3.45"]
tts = [
    "kokoro>=0.7.0",
    "soundfile>=0.12.0",
]
piper = ["piper-tts>=1.4.0"]
qwen = ["transformers>=4.40.0", "accelerate>=0.30.0", "torch>=2.0.0"]
fish = ["fish-speech>=1.5.0"]
diarize = ["pyannote.audio>=3.0"]
noise = ["noisereduce>=3.0"]
client = ["httpx>=0.27.0", "websockets>=12.0"]
all = [
    "vosk>=0.3.45",
    "piper-tts>=1.4.0",
    "kokoro>=0.7.0",
    "soundfile>=0.12.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/will-assistant/open-speech"
Repository = "https://github.com/will-assistant/open-speech"
Documentation = "https://github.com/will-assistant/open-speech#readme"
"Bug Tracker" = "https://github.com/will-assistant/open-speech/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
